tests/submission/Google/audio/autoplay/autoplay_reflects_autoplay.html

<!doctype html>
<html>
 <head>
  <title>audio.autoplay - reflection test</title>
  <script src="../../../../resources/testharness.js"></script>
  <script src="../../../../resources/testharnessreport.js"></script>
 </head>
 <body>
  <p><a href="http://dev.w3.org/html5/spec/Overview.html#dom-media-autoplay">spec reference</a></p>
  <audio id="audio" autoplay="autoplay">
  </audio>
  <div id="log"></div>
  <script>
test(function() {
 assert_true(
  document.getElementById("audio").autoplay,
  "audioElement.autoplay reflects true when present");
});
  </script>
 </body>
</html>
